fix: downloading wandb best model

+++ src/train.py
@@ -78,14 +78,17 @@ def train():
 
     file_name = "model-best.h5"
     if config["finetune"]:
-        os.remove("model-best.h5")
+        if os.path.exists(file_name):
+            os.remove("model-best.h5")
+
         api = wandb.Api()
         run = api.run(
             config["pretrained_model_link"]
         )  # different-sweep-34-efficientnet-epoch-3-val_f1_score-0.71.hdf5
         run.file(file_name).download()
         model = load_model(file_name)
-        print("Downloaded Trained model, finetuning...")
+        pml = config["pretrained_model_link"]
+        print(f"Downloaded Trained model: {pml},\nfinetuning...")
     else:
         # build model
         K.clear_session()
@@ -93,7 +96,7 @@ def train():
 
     # print(model.summary())
 
-    print(f"Model loaded: {model.name}\n\n")
+    print(f"Model loaded: {pml}.\n\n")
     opt = get_optimizer(config)
 
     config["metrics"].append(
